+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The major version bump is due to upping the required PHP version from `8.4` to `
 * Made improvements to the production error handlers:
 	- The 429 (too many requests) view will now tell the user when they can retry if a "retryAfter" date is provided.
 * Added CLI frame output component.
+* Added a `SimpleCache` (`PSR-16`) implementation to the framework core.
 
 #### Changes